+++ src/main/java/efub/SweetMeback/domain/post/dto/PostRequestDto.java
@@ -1,7 +1,6 @@
 package efub.SweetMeback.domain.post.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import efub.SweetMeback.domain.member.entity.Member;
 import efub.SweetMeback.domain.post.entity.Category;
 import efub.SweetMeback.domain.post.entity.Contact;
 import efub.SweetMeback.domain.post.entity.Meeting;
@@ -11,7 +10,7 @@ import java.time.LocalDateTime;
 
 @Getter
 public class PostRequestDto {
-    private Integer memberId;
+    private Long memberId;
     private String title;
     private String content;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
