+++ src/main/java/efub/SweetMeback/domain/post/repository/PostRepository.java
@@ -3,5 +3,7 @@ package efub.SweetMeback.domain.post.repository;
 import efub.SweetMeback.domain.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post,Integer> {
+import java.math.BigInteger;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
 }
